+++ src/main/java/me/postaddict/instagram/scraper/AnonymousInsta.java
@@ -1,5 +1,6 @@
 package me.postaddict.instagram.scraper;
 
+import me.postaddict.instagram.scraper.domain.Account;
 import me.postaddict.instagram.scraper.domain.Comment;
 import me.postaddict.instagram.scraper.domain.Media;
 
@@ -14,5 +15,6 @@ public interface AnonymousInsta extends StatelessInsta {
 
     List<Media> getTopMediasByTag(String tag) throws IOException;
     List<Comment> getCommentsByMediaCode(String code, int count) throws IOException;
+    Account getAccountById(long id) throws IOException;
 
 }
