+++ src/main/java/me/postaddict/instagram/scraper/AuthenticatedInsta.java
@@ -1,6 +1,5 @@
 package me.postaddict.instagram.scraper;
 
-import me.postaddict.instagram.scraper.domain.Account;
 import me.postaddict.instagram.scraper.domain.Comment;
 
 import java.io.IOException;
@@ -8,7 +7,6 @@ import java.io.IOException;
 public interface AuthenticatedInsta extends AnonymousInsta {
 
     void login(String username, String password) throws IOException;
-    Account getAccountById(long id) throws IOException;
     void likeMediaByCode(String code) throws IOException;
     void unlikeMediaByCode(String code) throws IOException;
